Refactor notifications, sidebar and add tests
-Refactor notifications to use template_tags so that the notifications do not need to be passed in at each view
-Refactor sidebar so it checks the current page automatically without requiring each page to set a bool manually
-Add tests for sidebar
-Add tests for shared_files view
-Add tests for notification model
-Add tests for notifications
-Add Javascript tests for the notification javascript
-Add test.html to run javascript tests

+++ tasks/views/notification_view.py
@@ -9,14 +9,21 @@ from django.forms.models import model_to_dict
 def process_notification_delete(request):
     """Processes a deletion of any notification"""
     notification_id = request.GET.get('notification_id')
-    #If the string passed in is delete-all, delete all notifications. 
-    #Otherwise, use the string as an id for a specific notification to delete
-    if notification_id=="delete-all":
-        Notification.objects.filter(user = request.user).delete()
-    elif notification_id is not None:
-        Notification.objects.filter(id=int(notification_id)).delete()
-    #Convert the notifications from a list of models to a dictionary so JSON can handle it
-    notifications = list(reversed(Notification.objects.filter(user = request.user)))
-    notifications = list(map(model_to_dict, notifications))
+    for_tests = request.GET.get('for_tests') #this is true if this view is being used for tests
+    if not for_tests:  
+        #When not testing, either delete all notifications, or a specific notification 
+        if notification_id=="delete-all":
+                #delete all the notifications of this user
+                Notification.objects.filter(user = request.user).delete()
+        elif notification_id is not None:
+            #delete the notification passed in
+            Notification.objects.filter(id=int(notification_id)).delete()
+        notifications = list(reversed(Notification.objects.filter(user = request.user)))
+        notifications = list(map(model_to_dict, notifications))
+    else:
+         #If testing, don't do anything to the database, and just return an empty list
+         #The purpose of this is to just show that this view is run 
+         notifications = []
+    #Return the data as a JsonResponse
     data = {'notifications': notifications}
     return JsonResponse(data)
